Fixed the stupid rendering issue where tiles would disappear near the edge of the screen. As I set the position of the images, the parameters were being cast to unsigned, so the values were skyrocketing to INT_MAX. I'm so stupid! TWO WEEKS!
Oh yeah... added some comments and documentation. Cleaned up some code.

+++ src/modes/map/map_objects.hpp
@@ -42,9 +42,9 @@ public:
     return draw_layer;
   }
 
-  void SetPosition(const float _x, const float _y);
+  void SetPosition(const int _x, const int _y);
 
-  void Move(const float _x, const float _y);
+  void Move(const int _x, const int _y);
 
   bool IsVisible();
